% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amonsieurx.R
\name{amonsieurx}
\alias{amonsieurx}
\title{Convert amonsieurx codes to numbers}
\usage{
amonsieurx(s, to_number = FALSE)
}
\arguments{
\item{s}{String with "amonsieurx" code}

\item{to_number}{Logical. Convert the result to numeric? If false, remains as character.}
}
\value{
Converted vector
}
\description{
Convert amonsieurx codes to numbers
}
