% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{compile_inches}
\alias{compile_inches}
\title{Given a two pairs of values and feet/inches, parse}
\usage{
compile_inches(d1, c1, d2, c2)
}
\arguments{
\item{d1}{Numeric.}

\item{d2}{Numeric}

\item{c2}{Character.}

\item{c2}{Character.}
}
\value{
Numeric.
}
\description{
If first is explicity feet, assume feet. Otherwise, assume inches. Note that,
because if() statements cannot be NA, we have to check cases in decreasing
order of NA-ness, so each must evaluate correctly and filter out any cases
that would break later statements.
}
