% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{general_dimension_extraction}
\alias{general_dimension_extraction}
\title{Create a tidy table of extracted dimensions}
\usage{
general_dimension_extraction(df, dimcol, idcol)
}
\value{
A data frame with an id column for joining to original table,
}
\description{
Create a tidy table of extracted dimensions
}
