% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{get_data}
\alias{get_data}
\title{Pull data from a given directory and name, and load into an object}
\usage{
get_data(dir, name)
}
\arguments{
\item{dir}{The directory to read from}

\item{name}{The name of the file (sans .rds)}
}
\value{
An R object stored in that data file
}
\description{
Pull data from a given directory and name, and load into an object
}
