% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{multi_join}
\alias{multi_join}
\title{Join a forign table multiple times, once for each reptition of a specified column}
\usage{
multi_join(df, jdf, from, to, n_reps, prefix)
}
\arguments{
\item{df}{Dataframe. The table on the left side of each join}

\item{jdf}{Dataframe. The table on the right side of each join}

\item{from}{Character. Quoted column name with the foreign key in the data frame to be joined}

\item{to}{Character. Quoted column name with the foreign key in the source data}

\item{n_reps}{Integer. Number of times `to` is repeated}
}
\value{
A dataframe
}
\description{
Join a forign table multiple times, once for each reptition of a specified column
}
