% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{multi_separate}
\alias{multi_separate}
\title{Repeatedly separate single fields into multiple ones}
\usage{
multi_separate(df, n_reps, into, ...)
}
\arguments{
\item{df}{Dataframe.}

\item{n_reps}{Integer. Number of times said column is repeated}

\item{into}{Character. Quoted names of columns (without any integer suffixes) to be produced after splitting}

\item{...}{Further arguments passed to separate_}
}
\value{
A dataframe.
}
\description{
This is useful when spreading apart a column that had been previously merged so as to be easier to repeatedly join using multi_join
}
