% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{norm_vars}
\alias{norm_vars}
\title{Wapper of unite/gather/separate to reshape groups of denormalized columns.}
\usage{
norm_vars(data, newcol = "group_no", base_names, n_reps, n_sep_char = "_",
  idcols, ...)
}
\arguments{
\item{data}{A data frame}

\item{newcol}{A temporary column name in which new values will be stored}

\item{base_names}{Quoted column names that are repeated as a unit. The
parameter \code{n_reps} will be treated as a suffix for uniting, gathering,
and separating these columns as groups.}

\item{n_reps}{Number of times the \code{base_names} repeat.}

\item{n_sep_char}{Delimiter that separates \code{base_names} from \code{n_reps}}

\item{idcols}{Quoted names of columns that should be kept as identifiers
when gathering these column groups.}
}
\description{
Specify the base names of the columns that are repeated (e.g. \code{c("a",
"b", "c")}), along with the number of times that those columns are repeated.
}
