% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{parse_dimensions}
\alias{parse_dimensions}
\title{Given a dimensions string, returns a tidy table with values, units, and dimension types.}
\usage{
parse_dimensions(s)
}
\arguments{
\item{s}{Character. Dimension strings to parse.}
}
\value{
A dataframe
}
\description{
Given a dimensions string, returns a tidy table with values, units, and dimension types.
}
