% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fractions.R
\name{parse_fraction}
\alias{parse_fraction}
\title{Parse fractions into numeric values}
\usage{
parse_fraction(s)
}
\arguments{
\item{s}{Character vector}
}
\value{
Numeric vector
}
\description{
Parse fractions into numeric values
}
