% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{parse_prices}
\alias{parse_prices}
\title{Generalized function for decimalizing and standardizing currencies on prices}
\usage{
parse_prices(source_dir, df, amount_col_name, currency_col_name, id_col_name,
  decimalized_col_name, aat_col_name, amonsieurx = FALSE,
  replace_slashes = FALSE)
}
\arguments{
\item{source_dir}{Directory from which to read currency authority tables.}

\item{df}{Source data frame}

\item{amount_col_name}{Quoted name of column with original price string.}

\item{currency_col_name}{Quoted name of column with original currency string.}

\item{id_col_name}{Quoted name of column with unique id.}

\item{decimalized_col_name}{Quoted name of column to produce containing
decimalized amount.}

\item{aat_col_name}{Quoted name of column to produce containing AAT code for
currency}

\item{amonsieurx}{Boolean. Apply \link{amonsieurx} transformation before
parsing?}

\item{replace_slashes}{Boolean. Replace all \code{/} characters with \code{-}
before parsing? (Necessary in cases where \code{/} is being used to
demarcate currency subunits, rather than fractions.)}
}
\value{
A data frame with 3 columns: \code{id_col_name} containing IDs from
  original \code{df}, and  new columns for decimalized amount and currency.
}
\description{
Generalized function for decimalizing and standardizing currencies on prices
}
