% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pick}
\alias{pick}
\title{Grab the first non-NA value of a vector}
\usage{
pick(v)
}
\arguments{
\item{v}{A vector}
}
\value{
A scalar
}
\description{
A useful function when calling \link{summarize}[dplyr]
}
