% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pipe_message}
\alias{pipe_message}
\title{Pipe-friendly message function}
\usage{
pipe_message(x, m)
}
\arguments{
\item{x}{An object}

\item{m}{Character. Message to be displayed using `message`}
}
\value{
x
}
\description{
Takes an object and a message, displays the message, and then returns the object. Ideal for use in a pipeline.
}
