% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{produce_artists_authority}
\alias{produce_artists_authority}
\title{Produce artist authority table from raw table.}
\usage{
produce_artists_authority(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce artist authority table from raw table.
}
