% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{produce_cpi}
\alias{produce_cpi}
\title{Produce tables with US consumer price index}
\usage{
produce_cpi(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce tables with US consumer price index
}
