% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{produce_currency_ids}
\alias{produce_currency_ids}
\title{Produce tables with currency AAT ids}
\usage{
produce_currency_ids(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce tables with currency AAT ids
}
