% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{produce_exchange_rates}
\alias{produce_exchange_rates}
\title{Produce tables with currency exchange rates}
\usage{
produce_exchange_rates(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce tables with currency exchange rates
}
