% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goupil.R
\name{produce_goupil}
\alias{produce_goupil}
\title{Produce Goupil table from raw table.}
\usage{
produce_goupil(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
This is a two-stage process. First, the raw, mostly denormalized exports are
read in, and then normalized into a series of tables. Secondly, IDs for
artists, buyers, sellers, objects, and transaction & inventory events are
generated and attached to the reamining core `goupil` table.
}
