% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knoedler.R
\name{produce_joined_knoedler}
\alias{produce_joined_knoedler}
\title{Produce a joined Knoedler table}
\usage{
produce_joined_knoedler(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Where to load preprocessed Knoedler files.}

\item{target_dir}{Where to save the fully joined Knoedler table.}
}
\value{
A data frame.
}
\description{
Produce a joined Knoedler table
}
