% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knoedler.R
\name{produce_knoedler}
\alias{produce_knoedler}
\title{Produce Knoedler table from raw table.}
\usage{
produce_knoedler(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce Knoedler table from raw table.
}
