% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knoedler.R
\name{produce_knoedler_stocknumber_concordance}
\alias{produce_knoedler_stocknumber_concordance}
\title{Editors have compiled links between stocknumbers that actually represent the
same object. We compose a graph of these relationships, identify connected
components that represnent all the stock numbers related to a single object,
and then produce a lookup table pairing each stock number with a UID for its
group}
\usage{
produce_knoedler_stocknumber_concordance(source_dir, target_dir)
}
\description{
Editors have compiled links between stocknumbers that actually represent the
same object. We compose a graph of these relationships, identify connected
components that represnent all the stock numbers related to a single object,
and then produce a lookup table pairing each stock number with a UID for its
group
}
