% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owners.R
\name{produce_owners_authority}
\alias{produce_owners_authority}
\title{Produce owners authority table from raw table.}
\usage{
produce_owners_authority(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce owners authority table from raw table.
}
