% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales_contents.R
\name{produce_sales_catalogs_info}
\alias{produce_sales_catalogs_info}
\title{Produce sales catalog info tables}
\usage{
produce_sales_catalogs_info(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce sales catalog info tables
}
