% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales_contents.R
\name{produce_sales_contents}
\alias{produce_sales_contents}
\title{Produce sales contents tables}
\usage{
produce_sales_contents(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce sales contents tables
}
