% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales_contents.R
\name{produce_sales_descriptions}
\alias{produce_sales_descriptions}
\title{Produce sales descriptions tables}
\usage{
produce_sales_descriptions(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{target_dir}{Path where resulting RDS files are saved}
}
\description{
Produce sales descriptions tables
}
