% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{produce_ulan_derivative_artists}
\alias{produce_ulan_derivative_artists}
\title{Produce derivative artist authority table for ULAN}
\usage{
produce_ulan_derivative_artists(source_dir, repo_path)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{repo_path}{Path where derivative CSV should be saved}
}
\description{
Produce derivative artist authority table for ULAN
}
