% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owners.R
\name{produce_ulan_derivative_owners}
\alias{produce_ulan_derivative_owners}
\title{Produce derivative owner authority table for ULAN}
\usage{
produce_ulan_derivative_owners(source_dir, repo_path)
}
\arguments{
\item{source_dir}{Path where source RDS files are found}

\item{repo_path}{Path where derivative CSV should be saved}
}
\description{
Produce derivative owner authority table for ULAN
}
