% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{produce_union_aat}
\alias{produce_union_aat}
\title{Collect all AAT ids used across GPI datasets and produce list of unique terms}
\usage{
produce_union_aat(source_dir)
}
\description{
Collect all AAT ids used across GPI datasets and produce list of unique terms
}
