% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{pull_star_exports}
\alias{pull_star_exports}
\title{Pull the latest exports from STAR into a temporary directory}
\usage{
pull_star_exports(out_dir, export_repo)
}
\value{
Path to temporary directory with the STAR exports
}
\description{
Pull the latest exports from STAR into a temporary directory
}
