% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_all_concordances}
\alias{read_all_concordances}
\title{Pull the latest Google Sheets concordances and parse into dataframes}
\usage{
read_all_concordances(out_dir, data_dict)
}
\arguments{
\item{out_dir}{Directory to which rds files ought to be written}

\item{data_dict}{Path to the data files definition YAML}
}
\description{
Pull the latest Google Sheets concordances and parse into dataframes
}
