% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_all_exports}
\alias{read_all_exports}
\title{Pull the latest STAR exports and parse into dataframes}
\usage{
read_all_exports(out_dir, data_dict, repo_path)
}
\arguments{
\item{out_dir}{Directory to which rds files ought to be written}

\item{data_dict}{Path to the data files definition YAML}

\item{repo_path}{Location of git repository holding GPI data}
}
\description{
The resulting files are stored in the `data-raw` directory. If any files have
more than 1% parsing errors, then an error will be thrown.
}
