% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rollup}
\alias{rollup}
\title{Collapse multiple values into one string with a separator}
\usage{
rollup(v, sep = "; ", collapse = TRUE)
}
\arguments{
\item{v}{A vector}

\item{sep}{Character to separate multiple values with}

\item{collapse}{Boolean. Only return unique values?}
}
\value{
A character scalar
}
\description{
A useful function when calling \link{summarize}[dplyr]
}
