% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{save_data}
\alias{save_data}
\title{Save an object to working data directory}
\usage{
save_data(dir, obj, objname = NULL)
}
\arguments{
\item{dir}{The directory to write to}

\item{obj}{The object to save}
}
\value{
Invisibly returns the path of the saved object
}
\description{
The object will be saved as dir/name_of_object.rds
}
