% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{single_separate}
\alias{single_separate}
\title{Detect how many columns to split a single col into}
\usage{
single_separate(df, source_col, sep = ";")
}
\arguments{
\item{df}{Data frame.}

\item{source_col}{Quoted name of source column}

\item{sep}{Character used as a delimiter within the column}
}
\value{
A data frame.
}
\description{
Detect how many columns to split a single col into
}
