% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{spread_out}
\alias{spread_out}
\title{Take a long table of multiple linked columns and reshape to a wide table}
\usage{
spread_out(df, idcol)
}
\arguments{
\item{df}{Data frame.}

\item{idcol}{Quoted name of a single id column.}
}
\value{
Data frame.
}
\description{
Take a long table of multiple linked columns and reshape to a wide table
}
