% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{pull_star_exports}
\alias{pull_star_exports}
\title{Pull the latest exports from STAR into a temporary directory}
\usage{
pull_star_exports(out_dir, export_repo)
}
\arguments{
\item{out_dir}{Destination directory for repo files}

\item{export_repo}{URL to git repository}
}
\value{
Path to temporary directory with the STAR exports
}
\description{
This will request a zip archive of the HEAD state of the repository, and
decompress those files into out_dir.
}
